+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CLI is renamed to `asynq`.
 - To upgrade the CLI to the latest version run `go get -u github.com/hibiken/tools/asynq`
 - The `ps` command in CLI is renamed to `servers`
+- `Concurrency` defaults to the number of CPUs when unset or set to a negative value.
 
 ### Added